# api/openapi.yaml
openapi: 3.1.0
info:
  title: Halo management API
  version: 0.1.0
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    The JSON API behind the Halo console and account portal. Anything the web interface does, a script can do with the same requests.

    **Authentication.** Send either the `halo_session` cookie of a signed-in person, or an API key of a service account as `Authorization: Bearer hlk_…`. API keys work only when they have the `api` scope, act with the roles of their service account, and are ignored on the personal routes under `/me` and on the sign-in routes under `/auth`. The personal routes under `/me/` (with a trailing segment, not `/me` itself) also accept an access token that `halo login` obtained for the built-in Halo CLI application.

    **Roles.** Each operation states the roles it requires. A global administrator passes every role check. "Any administrator" means security, user, helpdesk or application administrator, or auditor.

    **Errors.** Every error is `{"error": {"code": "ERR_…", "message": "…"}}`. Match on `code`; show `message` to people. Besides the codes listed per operation, any request can fail with `500` and `ERR_INTERNAL`, and unknown routes answer `404` with `ERR_NO_ROUTE`. Requests with a method other than GET, HEAD or OPTIONS that come from another website get `403` with `ERR_CROSS_SITE`.

...
servers:
  - url: /api/v1
security:
  - session: []
  - apiKey: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List every person
      description: 'Roles: any administrator. Service accounts are not included.'
      responses:
        '200':
          description: Every person, sorted by name.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Users]
      operationId: inviteUser
      summary: Invite a person
      description: 'Roles: user administrator. Creates the account with status `invited` and a setup link, and emails the link when SMTP is configured.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name]
              properties:
                email: {type: string, format: email, maxLength: 254}
                name: {type: string, maxLength: 200}
                title: {type: string, maxLength: 200}
                department: {type: string, maxLength: 200}
                location: {type: string, maxLength: 200}
                managerId: {type: string, description: Id of another person.}
                groupIds:
                  type: array
                  description: Assigned groups only.
                  items: {type: string}
      responses:
        '201':
          description: The new person and their setup link.
          content:
            application/json:
              schema:
                type: object
                required: [user, enrollUrl, expiresAt, emailed]
                properties:
                  user: {$ref: '#/components/schemas/User'}
                  enrollUrl: {type: string, format: uri}
                  expiresAt: {type: string, format: date-time}
                  emailed: {type: boolean, description: True when Halo also emailed the link.}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /users/{id}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Users]
      operationId: getUser
      summary: Get one person
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The person.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Users]
      operationId: updateUser
      summary: Change a person's profile
      description: 'Roles: user or helpdesk administrator. Helpdesk administrators can change title, department and location only. Only a global administrator can change a person who holds an administrator role. Send only the fields to change; an empty `managerId` removes the manager.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: {type: string, format: email}
                name: {type: string, maxLength: 200}
                title: {type: string, maxLength: 200}
                department: {type: string, maxLength: 200}
                location: {type: string, maxLength: 200}
                managerId: {type: string}
      responses:
        '200':
          description: The updated person.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /users/{id}/sessions:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Users]
      operationId: listUserSessions
      summary: List a person's active sessions
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The sessions.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Session'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /users/{id}/sign-ins:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Users]
      operationId: listUserSignIns
      summary: List a person's sign-in events
      description: 'Roles: any administrator. The 500 most recent.'
      responses:
        '200':
          description: The events, newest first.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/SignInEvent'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /users/{id}/suspend:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Users]
      operationId: suspendUser
      summary: Suspend a person and end their sessions
      description: 'Roles: user or security administrator. Refused for your own account and for the last active global administrator.'
      responses:
        '200': {$ref: '#/components/responses/UserResponse'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
  /users/{id}/restore:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Users]
      operationId: restoreUser
      summary: Restore a suspended person
      description: 'Roles: user administrator. The person becomes `active`, or `invited` when they never signed in and have no sign-in method.'
      responses:
        '200': {$ref: '#/components/responses/UserResponse'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /users/{id}/revoke-sessions:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Users]
      operationId: revokeUserSessions
      summary: End a person's sessions
      description: 'Roles: user, helpdesk or security administrator. When you target yourself, your current session stays.'
      responses:
        '200': {$ref: '#/components/responses/Revoked'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /users/{id}/reset-authentication:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Users]
      operationId: resetAuthentication
      summary: Remove every sign-in method and create a new setup link
      description: 'Roles: user or helpdesk administrator. Also ends every session and emails the link when SMTP is configured.'
      responses:
        '200':
          description: The new setup link.
          content:
            application/json:
              schema:
                type: object
                required: [enrollUrl, expiresAt, emailed]
                properties:
                  enrollUrl: {type: string, format: uri}
                  expiresAt: {type: string, format: date-time}
                  emailed: {type: boolean}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /users/{id}/methods/{methodId}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - name: methodId
        in: path
        required: true
        schema: {type: string}
    delete:
      tags: [Users]
      operationId: deleteUserMethod
      summary: Remove one of a person's sign-in methods
      description: 'Roles: user or helpdesk administrator.'
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /users/{id}/roles:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    put:
      tags: [Users]
      operationId: setUserRoles
      summary: Replace a person's administrator roles
      description: 'Roles: global administrator. Send the complete list; an empty list removes every role. You cannot remove your own global administrator role.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roles]
              properties:
                roles:
                  type: array
                  items: {$ref: '#/components/schemas/RoleKey'}
      responses:
        '200': {$ref: '#/components/responses/UserResponse'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
  /groups:
    get:
      tags: [Groups]
      operationId: listGroups
      summary: List every group with its member count
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The groups.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Group'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Groups]
      operationId: createGroup
      summary: Create a group
      description: 'Roles: user administrator.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kind]
              properties:
                name: {type: string, maxLength: 200}
                description: {type: string, maxLength: 2000}
                kind: {type: string, enum: [assigned, dynamic]}
                rule: {type: string, description: Required for dynamic groups., examples: ['user.department == "Engineering"']}
      responses:
        '201':
          description: The group.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Group'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /groups/{id}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Groups]
      operationId: getGroup
      summary: Get one group
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The group.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Group'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Groups]
      operationId: updateGroup
      summary: Rename a group, or change its description or rule
      description: 'Roles: user administrator. The kind cannot change.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: {type: string, maxLength: 200}
                description: {type: string, maxLength: 2000}
                kind: {type: string, enum: [assigned, dynamic], description: Accepted only when it equals the current kind.}
                rule: {type: string, description: Dynamic groups only.}
      responses:
        '200':
          description: The updated group.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Group'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
    delete:
      tags: [Groups]
      operationId: deleteGroup
      summary: Delete a group
      description: 'Roles: user administrator. Refused with `409`, naming each one, while an application, access package, policy, identity provider (for accounts created on first sign-in), open access review or lifecycle rule still uses the group. Completed access reviews are kept with the group''s name.'
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /groups/{id}/members:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Groups]
      operationId: listGroupMembers
      summary: List a group's members, or preview a rule
      description: 'Roles: any administrator. Deprovisioned people are left out.'
      parameters:
        - name: rule
          in: query
          description: Dynamic groups only. Returns who would match this rule instead of the saved one.
          schema: {type: string}
      responses:
        '200':
          description: The members.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
    post:
      tags: [Groups]
      operationId: addGroupMember
      summary: Add a person to an assigned group
      description: 'Roles: user administrator. Adding someone who is already a member succeeds.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: {type: string}
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/DynamicGroup'}
        '422': {$ref: '#/components/responses/Invalid'}
  /groups/{id}/members/{userId}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {$ref: '#/components/parameters/UserId'}
    delete:
      tags: [Groups]
      operationId: removeGroupMember
      summary: Remove a person from an assigned group
      description: 'Roles: user administrator.'
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/DynamicGroup'}
  /applications:
    get:
      tags: [Applications]
      operationId: listApplications
      summary: List every application
      description: 'Roles: any administrator. Includes credentials without secret values, assigned groups and sign-in statistics.'
      responses:
        '200':
          description: The applications.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Application'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Applications]
      operationId: createApplication
      summary: Register an application
      description: 'Roles: application administrator. OpenID Connect web applications and services get a client secret, returned once. SAML applications take their settings in `saml` and are always `web`.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, protocol, type]
              properties:
                name: {type: string, maxLength: 200}
                description: {type: string, maxLength: 2000}
                protocol: {type: string, enum: [oidc, oauth, saml]}
                type: {type: string, enum: [web, spa, native, service]}
                homepage: {type: string, format: uri}
                redirectUris:
                  type: array
                  description: Required for OpenID Connect applications other than services. https only, except http://localhost and http://127.0.0.1.
                  items: {type: string, format: uri}
                postLogoutUris:
                  type: array
                  items: {type: string, format: uri}
                scopes:
                  type: array
                  items: {type: string}
                groupIds:
                  type: array
                  items: {type: string}
                setupGuide: {type: string, enum: [grafana, forgejo, generic-oidc, kubernetes, generic-saml, aws-iam-identity-center, slack]}
                saml: {$ref: '#/components/schemas/SAMLInput'}
      responses:
        '201':
          description: The application and, for web applications and services, the client secret.
          content:
            application/json:
              schema:
                type: object
                required: [application, clientSecret]
                properties:
                  application: {$ref: '#/components/schemas/Application'}
                  clientSecret: {type: [string, 'null'], examples: [hls_…]}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /applications/{id}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Applications]
      operationId: getApplication
      summary: Get one application
      description: 'Roles: any administrator. SAML applications include a `saml` object.'
      responses:
        '200': {$ref: '#/components/responses/ApplicationResponse'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Applications]
      operationId: updateApplication
      summary: Change an application
      description: 'Roles: application administrator. Send only the fields to change. SAML applications accept only name, description, homepage and owner. OpenID Connect applications must keep the `openid` scope.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: {type: string, maxLength: 200}
                description: {type: string, maxLength: 2000}
                homepage: {type: string}
                redirectUris:
                  type: array
                  items: {type: string}
                postLogoutUris:
                  type: array
                  items: {type: string}
                scopes:
                  type: array
                  items: {type: string}
                tokenPolicy:
                  type: object
                  properties:
                    accessTokenTtl: {type: integer, minimum: 60, maximum: 86400}
                    idTokenTtl: {type: integer, minimum: 60, maximum: 86400, description: Services may use 0.}
                    refreshTokenTtl: {type: integer, minimum: 0, maximum: 7776000, description: 0 stops issuing refresh tokens.}
                    rotation: {type: boolean}
                owner: {type: string, description: Id of a person. An empty string removes the owner.}
      responses:
        '200': {$ref: '#/components/responses/ApplicationResponse'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
    delete:
      tags: [Applications]
      operationId: deleteApplication
      summary: Delete an application with its credentials and group assignments
      description: 'Roles: application administrator.'
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409':
          description: '`ERR_BUILT_IN`: the Halo CLI application cannot be deleted; disable it instead.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
  /applications/{id}/secrets:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Applications]
      operationId: rotateClientSecret
      summary: Create a new client secret
      description: 'Roles: application administrator. Every older secret stops working 24 hours later. The body is optional.'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                label: {type: string, maxLength: 100}
      responses:
        '201':
          description: The new credential and its secret, shown once.
          content:
            application/json:
              schema:
                type: object
                required: [credential, clientSecret]
                properties:
                  credential: {$ref: '#/components/schemas/Credential'}
                  clientSecret: {type: string}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
  /applications/{id}/secrets/{credentialId}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - name: credentialId
        in: path
        required: true
        schema: {type: string}
    delete:
      tags: [Applications]
      operationId: revokeClientSecret
      summary: Revoke a client secret immediately
      description: 'Roles: application administrator.'
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /applications/{id}/disable:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Applications]
      operationId: disableApplication
      summary: Stop sign-in and revoke every token issued to the application
      description: 'Roles: application administrator.'
      responses:
        '200': {$ref: '#/components/responses/ApplicationResponse'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /applications/{id}/enable:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Applications]
      operationId: enableApplication
      summary: Allow sign-in again
      description: 'Roles: application administrator.'
      responses:
        '200': {$ref: '#/components/responses/ApplicationResponse'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /applications/{id}/groups:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    put:
      tags: [Applications]
      operationId: setApplicationGroups
      summary: Replace the groups assigned to an application
      description: 'Roles: application administrator. Send the complete list; an empty list removes every assignment.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupIds]
              properties:
                groupIds:
                  type: array
                  items: {type: string}
      responses:
        '200': {$ref: '#/components/responses/ApplicationResponse'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
  /applications/{id}/saml:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    put:
      tags: [Applications]
      operationId: setApplicationSAML
      summary: Change a SAML application's settings
      description: 'Roles: application administrator. Send any of the fields; `metadataXml` imports the entity ID and ACS URLs from the service provider''s metadata.'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SAMLInput'}
      responses:
        '200': {$ref: '#/components/responses/ApplicationResponse'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /sessions:
    get:
      tags: [Activity]
      operationId: listSessions
      summary: List every active session
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The sessions.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Session'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /signing-keys:
    get:
      tags: [Settings]
      operationId: listSigningKeys
      summary: List the OpenID Connect signing keys and SAML certificates
      description: 'Roles: any administrator.'
      responses:
        '200': {$ref: '#/components/responses/SigningKeysResponse'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: halo_session
      description: The session cookie Halo sets when a person signs in.
    apiKey:
      type: http
      scheme: bearer
      description: An API key of a service account, starting with `hlk_`, with the `api` scope. Requests act with the service account's roles.
    cliAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An access token that `halo login` obtained for the built-in Halo CLI application. Accepted only on routes under `/me/`.
...